# api-schemas/openapi.yaml
openapi: 3.1.1
info:
  title: RomBet
  version: 1.0.0
servers:
  - url: //dummy.com/
    description: Local development server
paths:
  /api/v1/teams:
    get:
      summary: Get all teams
      description: Returns a list of all teams
      operationId: getAllTeams
      parameters:
        - name: X-Forwarded-For
          in: header
          description: IP address of the client (optional for testing, may be automatically detected)
          required: false
          schema:
            type: string
            format: ipv4
      tags:
        - Team
      responses:
        '200':
          description: List of teams
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      example: "team_123"
                    name:
                      type: string
                      example: "Team A"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to fetch teams"
                  code:
                    type: integer
                    example: 500

    post:
      summary: Create a new team
      description: Creates a new team with the provided data
      operationId: createTeam
      parameters:
        - name: X-Forwarded-For
          in: header
          description: IP address of the client (optional for testing, may be automatically detected)
          required: false
          schema:
            type: string
            format: ipv4
      tags:
        - Team
      requestBody:
        description: Team data
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Team A"
              required: [name]
      responses:
        '201':
          description: Team created successfully
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Name is required"
                  code:
                    type: integer
                    example: 400
        '422':
          description: This team already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "This team already exists"
                  code:
                    type: integer
                    example: 422
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to create team"
                  code:
                    type: integer
                    example: 500
...